# txnkv: optimize batch-get by reducing overhead of backoffer