| Workload | Threads | QPS (this PR) | QPS (baseline) | Diff |
| -------- | ------: | ------------: | -------------: | ---: |
| batchget |      64 |      22437.15 |       21908.98 | 2.4% |

| Workload | Threads |  QPS  | P99 Lat (this PR) | P99 Lat (baseline) | Diff  |
| -------- | ------: | ----: | ----------------: | -----------------: | ----: |
| batchget |      64 | 16000 |            6.67ms |             6.91ms | -3.5% |